fix: rework relationship between collapse/expand and listener start/stop
The way collapse/expand was handled, we were actually always connecting
to all interface objects. All ports, all properties from the beginning
until the very end. This was caused by how the code handled the fact
that when a root is collapsed, its children aren't automatically
(but then, they are hidden and we want to stop reading their data)

This commit splits the collapse/expand management and the start/stop
to reach the expected behavior:
- we connect only when needed. For now, it means the first time the
  item category is opened (because otherwise the inspector shows
  "no data")
- readers are disconnected once the port is hidden from view (e.g.
  one of its parents is collapsed)
- works the same for properties