fix: do not create two listeners for port and property display
The code reuses the same item for both columns of the tree view ...
which is rather obviously wrong, but it is what it is.

This makes both items create readers. The second reader is removed
because of the (also wrong) existence of a global reader deep down
in the async code, but it costs at the beginning (admittedly, a lot
less now that we're not creating a reader for every single port)